Feature/services

+++ Homify.WebApi/Controllers/Admins/AdminController.cs
@@ -2,23 +2,30 @@ using Homify.BusinessLogic.Roles;
 using Homify.BusinessLogic.Users;
 using Homify.BusinessLogic.Users.Entities;
 using Homify.Exceptions;
+using Homify.Utility;
 using Homify.WebApi.Controllers.Admins.Models;
+using Homify.WebApi.Filters;
 using Microsoft.AspNetCore.Mvc;
+using Constants = Homify.Utility.Constants;
 
 namespace Homify.WebApi.Controllers.Admins;
 
 [ApiController]
 [Route("admins")]
-public sealed class AdminController : ControllerBase
+public sealed class AdminController : HomifyControllerBase
 {
     private readonly IUserService _userService;
+    private readonly IRoleService _roleService;
 
-    public AdminController(IUserService userService)
+    public AdminController(IUserService userService, IRoleService roleService)
     {
         _userService = userService;
+        _roleService = roleService;
     }
 
     [HttpPost]
+    [AuthenticationFilter]
+    [AuthorizationFilter(PermissionsGenerator.CreateAdmin)]
     public CreateAdminResponse Create(CreateAdminRequest? request)
     {
         if (request == null)
@@ -26,19 +33,23 @@ public sealed class AdminController : ControllerBase
             throw new NullRequestException("Request cannot be null");
         }
 
+        var adminRole = _roleService.GetRole(Constants.ADMINISTRATOR);
+
         var arguments = new CreateUserArgs(
             request.Name ?? string.Empty,
             request.Email ?? string.Empty,
             request.Password ?? string.Empty,
             request.LastName ?? string.Empty,
-            RolesGenerator.Admin());
+            adminRole);
 
-        var administratorSaved = _userService.AddUser(arguments);
+        var administratorSaved = _userService.AddAdmin(arguments);
 
         return new CreateAdminResponse(administratorSaved);
     }
 
     [HttpDelete("{adminId}")]
+    [AuthenticationFilter]
+    [AuthorizationFilter(PermissionsGenerator.DeleteAdmin)]
     public void Delete(string adminId)
     {
         var admin = _userService.GetById(adminId);
@@ -46,14 +57,20 @@ public sealed class AdminController : ControllerBase
         {
             throw new NotFoundException("Admin not found");
         }
-        else
+
+        if (admin.Role.Name != Constants.ADMINISTRATOR)
         {
-            _userService.Delete(adminId);
+            throw new InvalidOperationException("Target user is not an admin");
         }
+
+        _userService.Delete(adminId);
     }
 
-    [HttpGet]
-    public List<UserBasicInfo> AllAccounts([FromQuery] string limit, [FromQuery] string offset)
+    [HttpGet("accounts")]
+    [AuthenticationFilter]
+    [AuthorizationFilter(PermissionsGenerator.GetAllAccounts)]
+    public List<UserBasicInfo> AllAccounts([FromQuery] string? limit, [FromQuery] string? offset,
+        [FromQuery] string? role, [FromQuery] string? fullName)
     {
         var pageSize = 10;
         var pageOffset = 0;
@@ -69,6 +86,18 @@ public sealed class AdminController : ControllerBase
         }
 
         List<User> list = _userService.GetAll();
+
+        if (!string.IsNullOrEmpty(role))
+        {
+            list = list.Where(u => u.Role.Name.Contains(role, StringComparison.OrdinalIgnoreCase)).ToList();
+        }
+
+        if (!string.IsNullOrEmpty(fullName))
+        {
+            list = list.Where(u => Helpers.GetUserFullName(u.Name, u.LastName)
+                .Contains(fullName, StringComparison.OrdinalIgnoreCase)).ToList();
+        }
+
         var paginatedList = list.Skip(pageOffset).Take(pageSize).ToList();
 
         List<UserBasicInfo> result = [];
